Add disk_allocation_ratio attribute
This commit adds the disk_allocation_ratio attribute to the Nova
config. This allows Nova to schedule instances on hosts only if
there is sufficient disk space available for root and ephemeral
storage.

Note that this attribute is only enabled when DiskFilter is added
to node["nova"]["platform"]["nova_scheduler_default_filters"].