Allow override for vncserver_listen
Migrations fail when vncserver_listen != 0.0.0.0 because the libvirt
XML config for that VM contains the vncserver_listen address. When an
instance is migrated to a new host, VNC tries to bind to that interface,
which fails miserably.

This commit allows the operator to override the vncserver_listen config
by setting nova/libvirt/vncserver_listen. If this value is not defined,
it will default to the current lookup method using search.

Fixes rcbops/chef-cookbooks#399