Fix for issue 915
The xvpvncproxy is not equally supported on RHEL and Ubuntu alike,
so this fixes the issues in our cookbooks with the missing package
in RHEL by removing it from our cookbooks.  It no longer makes sense
to support the installation and monitoring of the installation of
xvpvnc when it the package is not supported on both RHEL and Ubuntu.

Issue:
https://github.com/rcbops/chef-cookbooks/issues/915