Fix Issue 904
This commit adds the ability for a user to set the VNC password.
If the password is not set one will be set for you. This is done
to address a potential security issue with not having password
protected VNC access by default.

Issue:
https://github.com/rcbops/chef-cookbooks/issues/904
(cherry picked from commit 6f41a4648a55e68ca1af175c650bc06e34644285)